sample_app/requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
black==19.10b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
coverage==5.5
coveralls==3.2.0
Django==3.1.8
django-filter==2.4.0
django-graphql-jwt==0.3.4
djangorestframework==3.12.4
docopt==0.6.2
entrypoints==0.3
flake8==3.7.9
flake8-black==0.1.1
flake8-bugbear==20.1.4
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.2
iniconfig==1.1.1
mccabe==0.6.1
mock==4.0.3
packaging==21.0
pathspec==0.9.0
pluggy==0.13.1
promise==2.3
psycopg2-binary==2.9.1
ptvsd==4.3.2
py==1.10.0
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
pytest-django==4.4.0
pytest-random-order==1.0.4
pytz==2021.1
regex==2021.8.21
requests==2.26.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.3
urllib3==1.26.6
